feat: Refactor thread support by relocating core/thread module to integration/thread and updating import paths
